+++ models/Session.py
@@ -6,9 +6,9 @@ from models.Device import Device
 class Session(Base, Model):
     __tablename__ = 'session'
 
-    id = Column(Integer, primary_key = True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey(User.id), nullable=False)
-    device_id = Column(Integer, ForeignKey(Device.id), nullable=False)
+    id = Column(BigInteger, primary_key = True, autoincrement=True)
+    user_id = Column(BigInteger, ForeignKey(User.id), nullable=False)
+    device_id = Column(BigInteger, ForeignKey(Device.id), nullable=False)
     token = Column(String(120), nullable=False)
     created = Column(DateTime, default = Utils.time())
     updated = Column(DateTime, default = Utils.time(), onupdate =  Utils.time())
